requirements.txt
ultralytics
opencv-python
pillow
numpy
pandas

torch
torchvision

streamlit

openai
python-dotenv

pyyaml

matplotlib
seaborn

tqdm
loguru
langchain
langchain-openai
langgraph
langchain-community
openai
ollama
ultralytics
python-dotenv
pyyaml
numpy
torch
torchvision

langchain
langchain-community
pypdf
langchain
langchain-community
sentence-transformers
pypdf
faiss-cpu

langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf